pyproject.toml
[project]
name = "shuffleupagus"
version = "1.0.0"
description = "generate and synchronize smart, balanced playlists"
requires-python = ">= 3.10"

dependencies = [
  "colorama           >= 0.4.6",
  "coloredlogs        >= 15.0",
  "joblib             >= 1.4.2",
  "pyyaml             >= 6.0.2",

  "apple-music-python >= 1.0.6",
  "py-applescript     >= 1.0.3",
  "spotipy            >= 2.25",
]

[project.scripts]
shuffleupagus = "shuffleupagus.shuffleupagus:main"

[tool.black]
line-length = 120
